requirements.txt
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
idna==3.3
psycopg2==2.9.3
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
